Python - Modo 0/requirements.txt
python_vlc==3.0.11115
python_osc==1.7.4
